pyproject.toml
[project]
name = "nba-underdog-bet"
version = "0.5.0"
description = "AI-powered NBA underdog betting analysis system"
requires-python = ">=3.11"
dependencies = [
    "pydantic-ai>=0.0.15",
    "pydantic-settings>=2.2.0",
    "httpx>=0.27.0",
    "pandas>=2.2.0",
    "python-dotenv>=1.0.0",
    "rich>=13.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.3.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
